Moving some files from bigtable-hbase to bigtable-grpc-interface
I had to add a setter for userAgent in BigtableOptions, since BigtableConstants.USER_AGENT is specific to bigtable-hbase.
