+++ bigtable-grpc-interface/src/main/java/com/google/cloud/bigtable/config/Logger.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.bigtable.hbase;
+package com.google.cloud.bigtable.config;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
