+++ bigtable-hbase/src/main/java/com/google/cloud/bigtable/hbase/adapters/ColumnFamilyFormatter.java
@@ -17,7 +17,7 @@ package com.google.cloud.bigtable.hbase.adapters;
 
 import org.apache.hadoop.hbase.TableName;
 
-import com.google.cloud.bigtable.hbase.BigtableOptions;
+import com.google.cloud.bigtable.config.BigtableOptions;
 import com.google.common.base.Preconditions;
 
 public class ColumnFamilyFormatter {
