+++ bigtable-hbase/src/main/java/com/google/cloud/bigtable/hbase/adapters/SampledRowKeysAdapter.java
@@ -16,7 +16,7 @@
 package com.google.cloud.bigtable.hbase.adapters;
 
 import com.google.bigtable.v1.SampleRowKeysResponse;
-import com.google.cloud.bigtable.hbase.Logger;
+import com.google.cloud.bigtable.config.Logger;
 
 import org.apache.hadoop.hbase.HConstants;
 import org.apache.hadoop.hbase.HRegionInfo;
