+++ bigtable-hbase/src/main/java/com/google/cloud/bigtable/hbase/adapters/TableAdapter.java
@@ -17,7 +17,7 @@ package com.google.cloud.bigtable.hbase.adapters;
 
 import com.google.bigtable.admin.table.v1.ColumnFamily;
 import com.google.bigtable.admin.table.v1.Table;
-import com.google.cloud.bigtable.hbase.BigtableOptions;
+import com.google.cloud.bigtable.config.BigtableOptions;
 
 import org.apache.hadoop.hbase.HColumnDescriptor;
 import org.apache.hadoop.hbase.HTableDescriptor;
